+++ src/main/java/com/fasterxml/jackson/databind/deser/impl/PropertyValueBuffer.java
@@ -335,7 +335,12 @@ public class PropertyValueBuffer
                 // also: may need to set a property value as well
                 SettableBeanProperty idProp = _objectIdReader.idProperty;
                 if (idProp != null) {
-                    return idProp.setAndReturn(bean, _idValue);
+                    // [databind#5328] Records do not have setters, skip...to set id value
+                    if (bean.getClass().isRecord()) {
+                        return bean;
+                    } else  {
+                        return idProp.setAndReturn(bean, _idValue);
+                    }
                 }
             } else {
                 // 07-Jun-2016, tatu: Trying to improve error messaging here...
